+++ src/Trans.tsx
@@ -1,12 +1,15 @@
 import parse from 'html-react-parser';
 
 interface TransProps {
-  translation: string;
+  translation: string | string[];
   children: React.ReactNode;
 }
 
 const Trans = ({ translation, children }: TransProps): any => {
-  return translation ? parse(translation) : children;
+  const calculateTranslation =
+    translation instanceof Array ? translation.join(' ') : translation;
+
+  return calculateTranslation ? parse(calculateTranslation) : children;
 };
 
 export default Trans;